# perf: rework benchmarks