--- a/.eslintignore
+++ b/.eslintignore
@@ -2,4 +2,3 @@
 calc
 docs
 e2e-tests
-scripts