--- a/scripts/benchmarkConfig.ts
+++ b/scripts/benchmarkConfig.ts
@@ -10,40 +10,40 @@ export const benchmarkConfig: IBenchmarkConfig = {
     '/accounts/{accountId}/staking-info': {
         path: '/benchmarks/accountsStakingInfo',
     },
-    // '/accounts/{accountId}/staking-payouts': {
-    //     path: '/benchmarks/accountsStakingPayouts',
-    // },
-    // '/accounts/{accountId}/validate': {
-    //     path: '/benchmarks/accountsValidate',
-    // },
-    // '/blocks/{blockId}': {
-    //     path: '/benchmarks/blocks',
-    // },
-    // '/pallets/staking/progress': {
-    //     path: '/benchmarks/palletsStakingProgress',
-    // },
-    // '/pallets/{palletId}/storage': {
-    //     path: '/benchmarks/palletsPalletIdStorage',
-    // },
-    // '/pallets/{palletId}/storage/{storageItemId}': {
-    //     path: '/benchmarks/pallets/palletsPalletIdStorageStorageId',
-    // },
-    // '/paras': {
-    //     path: '/benchmarks/paras',
-    // },
-    // '/paras/leases/current': {
-    //     path: '/benchmarks/parasLeasesCurrent',
-    // },
-    // '/paras/auctions/current': {
-    //     path: '/benchmarks/parasAuctionCurrent',
-    // },
-    // '/paras/crowdloans': {
-    //     path: '/benchmarks/parasCrowdloans',
-    // },
-    // '/paras/{paraId}/crowdloan-info': {
-    //     path: '/benchmarks/parasParaIdCrowdloanInfo',
-    // },
-    // '/paras/{paraId}/lease-info': {
-    //     path: '/benchmarks/parasParaIdLeasesInfo',
-    // },
+    '/accounts/{accountId}/staking-payouts': {
+        path: '/benchmarks/accountsStakingPayouts',
+    },
+    '/accounts/{accountId}/validate': {
+        path: '/benchmarks/accountsValidate',
+    },
+    '/blocks/{blockId}': {
+        path: '/benchmarks/blocks',
+    },
+    '/pallets/staking/progress': {
+        path: '/benchmarks/palletsStakingProgress',
+    },
+    '/pallets/{palletId}/storage': {
+        path: '/benchmarks/palletsPalletIdStorage',
+    },
+    '/pallets/{palletId}/storage/{storageItemId}': {
+        path: '/benchmarks/pallets/palletsPalletIdStorageStorageId',
+    },
+    '/paras': {
+        path: '/benchmarks/paras',
+    },
+    '/paras/leases/current': {
+        path: '/benchmarks/parasLeasesCurrent',
+    },
+    '/paras/auctions/current': {
+        path: '/benchmarks/parasAuctionCurrent',
+    },
+    '/paras/crowdloans': {
+        path: '/benchmarks/parasCrowdloans',
+    },
+    '/paras/{paraId}/crowdloan-info': {
+        path: '/benchmarks/parasParaIdCrowdloanInfo',
+    },
+    '/paras/{paraId}/lease-info': {
+        path: '/benchmarks/parasParaIdLeasesInfo',
+    },
 }