--- a/scripts/benchmarkConfig.ts
+++ b/scripts/benchmarkConfig.ts
@@ -63,5 +63,8 @@ export const benchmarkConfig: IBenchmarkConfig = {
     },
     '/node/network': {
         path: '/benchmarks/nodeNetwork',
+    },
+    '/node/transaction-pool': {
+        path: '/benchmarks/nodeTransactionPool',
     }
 }