--- a/scripts/benchmarkConfig.ts
+++ b/scripts/benchmarkConfig.ts
@@ -66,5 +66,8 @@ export const benchmarkConfig: IBenchmarkConfig = {
     },
     '/node/transaction-pool': {
         path: '/benchmarks/nodeTransactionPool',
-    }
+    },
+    '/node/version': {
+        path: '/benchmarks/nodeVersion',
+    },
 }