--- a/scripts/config.ts
+++ b/scripts/config.ts
@@ -80,8 +80,8 @@ export const historicalE2eConfig: Record<string, IChainConfigE2E> = {
 			...defaultJestOpts,
 			args: ['start:historical-e2e-tests', '--chain', 'statemint'],
 		},
-		SasStartOpts: defaultSasStartOpts
-	}
+		SasStartOpts: defaultSasStartOpts,
+	},
 };
 
 export const latestE2eConfig: Record<string, IChainConfigE2E> = {
@@ -92,8 +92,8 @@ export const latestE2eConfig: Record<string, IChainConfigE2E> = {
 			proc: 'latest-e2e',
 			resolver: 'Finished with a status code of 0',
 			resolverFailed: 'Finished with a status code of 1',
-			args: ['start:latest-e2e-tests', '--chain', 'polkadot']
-		}
+			args: ['start:latest-e2e-tests', '--chain', 'polkadot'],
+		},
 	},
 	statemint: {
 		wsUrl: 'wss://statemint-rpc.polkadot.io',
@@ -102,7 +102,7 @@ export const latestE2eConfig: Record<string, IChainConfigE2E> = {
 			proc: 'latest-e2e',
 			resolver: 'Finished with a status code of 0',
 			resolverFailed: 'Finished with a status code of 1',
-			args: ['start:latest-e2e-tests', '--chain', 'statemint']
-		}
-	}
-}
+			args: ['start:latest-e2e-tests', '--chain', 'statemint'],
+		},
+	},
+};