# Sidecar Benchmarks

Sidecar's benchmarks utilize `lua`, `wrk`, and an archive node (local). This PR aims at completely a few things:

- Providing benchmarks for every available endpoint that can be queried against a `polkadot` archive node.
- Writing a script that would allow simple access to benchmarking for either one or all endpoints.
- Having the script write the results to a `benchmarks.txt` file in the root of the repository.
- Re-activate eslint for the scripts folder (This is what makes the PR seem a little bigger than it should be).

## The script

This script's purpose is to run the benchmarks inside of the `<root>/benchmarks` directory. This script requires you to have `wrk`, `lua`, and a fully synced archive node. Please reference `<root>/benchmarks/README.md` for more information on installation. 

In order to start this script locally, run from the root directory of this repository:

ex1:
```bash
$ yarn
$ yarn bench
```

ex2:
```bash
$ yarn
$ yarn bench --log-level=info --time=30s  
```

ex3:
```
$ yarn
$ yarn bench --log-level=info --time=30s  --endpoint=/accounts/{accountId}/balance-info
```

### Flags

`--ws-url`: The Ws Url to run the benchmarks against. Default is `ws://127.0.0.1:9944`.

`--endpoint`: Run a single benchmark by specificing which endpoint to run. If this is absent it will run all benchmarks.

`--log-level`: The log-level to run the benchmarks in. Defaults to `http`.

`--time`: The amount of time each benchmark should run for. Ex: `1m`, `30s`, `15s`. Default is `1m`.

- [x] Benches for all endpoints.
    - [X] `/accounts/{accountId}/balance-info`
    - [x] `/accounts/{accountId}/vesting-info`
    - [x] `/accounts/{accountId}/validate`
    - [x]  `/accounts/{accountId}/convert`
    - [x] `/accounts/{accountId}/staking-info`
    - [x] `/accounts/{accountId}/staking-payouts`
    - [X] `/blocks/{blockId}`
    - [x] `/pallets/staking/progress`
    - [x] `/pallets/{palletId}/storage`
    - [x] `/pallets/{palletId}/storage/{storageItemId}`
    - [x] `/pallets/nomination-pools/info`
    - [x] `/pallets/nomination-pools/{poolId}`
    - [x] `/pallets/{palletId}/errors`
    - [x] `/pallets/{palletId}/errors/{errorItemId}`
    - [x] `/pallets/staking/validators`
    - [x] `/transaction/material`
    - [x] `/runtime/spec`
    - [x] `/paras`
    - [x] `/paras/leases/current`
    - [x] `/paras/auctions/current`
    - [x] `/paras/crowdloans`
    - [x] `/paras/{paraId}/crowdloan-info`
    - [x] `/paras/{paraId}/lease-info`
    - [x] `/blocks/{blockId}/header`
    - [x] `/blocks/{blockId}/extrinsics/{extrinsicIndex}`
    - [x] `/blocks/head`
    - [x] `/blocks/head/header`
    - [x] `/node/network`
    - [x] `/node/transaction-pool`
    - [x] `/node/version`
- [x] flushed out Lua util module
- [x] README is finished with instructions on running locally in both linux and mac
- [x] Script to run one or all benchmarks

closes: https://github.com/paritytech/substrate-api-sidecar/issues/1050