@TarikGul 2022-09-06
@radupopa2010 Hey just wanted to give you a heads up as well, I am completely reworking the benchmarks in this PR. I will ping you when I work through the CI portion which should be sometime this week when I get the rest finished

cc: @alvicsam