@IkerAlus 2023-02-08
Couple of questions before actually running this:

- The performance of certain endpoints are highly dependent on the current (or at specific timepoint) status of the network (f.e., transaction pool may be empty or full). Will this be somehow captured in the benchmark report?
- Could this be run being connected to a public endpoint node and/or a non-archive node (assuming we omit those endpoints that query the blockchain in the past?