@TarikGul 2023-02-08
> Couple of questions before actually running this:
> 
> * The performance of certain endpoints are highly dependent on the current (or at specific timepoint) status of the network (f.e., transaction pool may be empty or full). Will this be somehow captured in the benchmark report?
> * Could this be run being connected to a public endpoint node and/or a non-archive node (assuming we omit those endpoints that query the blockchain in the past?

> Will this be somehow captured in the benchmark report?

So no because we should only be querying a local archive node, and not a public node. Most of the benchmarks work with historic blocks, there are some exceptions though. If the transaction-pool is full or not doesn't have a huge impact in my experience from running the benchmarks for a while now (But if it does it'll have a small impact as the amount of queries are so large that it will be a small outlier). 

> Could this be run being connected to a public endpoint node and/or a non-archive node (assuming we omit those endpoints that query the blockchain in the past?

As mentioned above too we definitely don't want to benchmark against public live nodes because of the stress the benchmarks would impose upon the network.